#195: fix the tests for watch movies
There is an old error here from when I split the movies and watchedMovies query. This was redefining the movies query with the watchedMovies response and somehow everything was passing. Now that the code throws if movies or lists has errors, the error that has been there for a while is surfacing.